% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{age_in_years}
\alias{age_in_years}
\title{Compute complete years between events.}
\usage{
age_in_years(begin_date, end_date)
}
\arguments{
\item{begin_date}{Beginning date (usually birth).}

\item{end_date}{Ending date (usually death).}
}
\value{
Number of anniversaries of begin_date (usually person's age).
}
\description{
Compute complete years between events.
}
\examples{
# Dave Brubeck
age_in_years("1920-12-06", "2012-12-05")
# John F. Kennedy
age_in_years("1917-05-29", "1963-11-22")
# Elvis Presley
age_in_years("1935-01-08", "1977-08-16")
# Groucho Marx
age_in_years("1890-10-02", "1977-08-19")
}
