% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Cancer indidence data.}
\format{A data table with 646 rows and 5 variables.
\describe{
  \item{\code{Year}}{Calendar year}
  \item{\code{agegroup}}{SEER age group 0, 1-4, 5-9, 10-14, ... 75-84, 85+}
  \item{\code{Sex}}{Male or Female}
  \item{\code{n}}{Number of newly diagnosed cases}
  \item{\code{pop}}{Census estimated population}
 }}
\source{
{CDC WONDER online database} \url{https://wonder.cdc.gov/}
}
\usage{
data(cancer)
}
\description{
Dataset of cancer incidence by year, sex and age group, United States 1999-2015
}
\references{
United States Cancer Statistics (1999-2015) Incidence, WONDER
 Online Database. United States Department of Health and Human Services,
 Centers for Disease Control and Prevention and National Cancer Institute.
 http://wonder.cdc.gov/cancer-v2015.html
}
\keyword{datasets}
