% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated-ratio.R
\name{correlated_rates}
\alias{correlated_rates}
\title{Age-adjusted rates for correlated variables}
\usage{
correlated_rates(df, region, agegroup, events, person_yrs, std_pop,
  parent = "Parent Region", base = 1e+05, level = 95, dec_rate = 1,
  dec_ratio = 2)
}
\arguments{
\item{df}{A data frame with columns for age group, event counts, and person-years totals as described in the next three arguments}

\item{region}{Subregion for which to calculate adjusted rates; e.g. county within a state.}

\item{agegroup}{Age group or other stratifying variable.}

\item{events}{Number of events.}

\item{person_yrs}{Number of person-years at risk.}

\item{std_pop}{Vector of standard population distribution. Can be totals, proportions, or percentages.}

\item{parent}{Name of parent region.}

\item{base}{Multiplier; e.g. per 100,000 population.}

\item{level}{Confidence level expressed as percentage.}

\item{dec_rate}{Decimal places to round adjusted rates.}

\item{dec_ratio}{Decimal places to round rate ratios.}
}
\value{
A data table with the following fields:
\describe{
  \item{\code{events}}{Number of events.}
  \item{\code{person_yrs}}{Total person-years at risk.}
  \item{\code{adj_rate}}{Age-adjusted rate.}
  \item{\code{adj_rate_stderr}}{Standard error of age-adjusted rate.}
  \item{\code{adj_lci}}{Lower confidence limit of age-adjusted rate per Tiwari (2006)}
  \item{\code{adj_uci}}{Upper confidence limit for age-adusted rate per Tiwari (2006)}
  }
}
\description{
Age-adjusted rates for correlated variables
}
\references{
Tiwari RC et al. (2006) Efficient interval estimation for age-adjusted 
cancer rates. Statistical Methods in Medical Research 15:547-569. 
https://www.ncbi.nlm.nih.gov/pubmed/17260923

Tiwari RC et al. (2010) Interval estimation for ratios of correlated age-adjusted rates.
Journal of Data Science 8:471-482. https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3279758/
}
