% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crude_rate.R
\name{crude_rate}
\alias{crude_rate}
\title{Crude or age-specific rates.}
\usage{
crude_rate(df, events, person_yrs, base = 1e+05, level = 95, decimals = 1)
}
\arguments{
\item{df}{A data frame.}

\item{events}{Number of events.}

\item{person_yrs}{Number of person-years at risk.}

\item{base}{Multiplier; e.g. per 100,000 population.}

\item{level}{Confidence level expressed as percentage.}

\item{decimals}{Decimal places to round results.}
}
\value{
A data table with the following fields:
\describe{
  \item{\code{events}}{Number of events.}
  \item{\code{person_yrs}}{Total person-years at risk.}
  \item{\code{rate}}{Crude (unadjusted) rate.}
  \item{\code{rate_stderr}}{Standard error of rate.}
  \item{\code{rate_lci}}{Lower confidence limit for rate.}
  \item{\code{rate_uci}}{Upper confidence limit for rate.}
  }
}
\description{
Crude or age-specific rates.
}
\note{
Confidence limits for crude rates are computed using the method of 
Garwood (1936).
}
\examples{
# US cancer rates by year
library(dplyr)
cancer \%>\% 
  group_by(Year) \%>\% 
  summarize_at(c("n", "pop"), sum) \%>\% 
  do(crude_rate(., n, pop))

}
\references{
Garwood F (1936) Fiducial limits for the Poisson distribution,
 Biometrika 28:437-442.
}
