% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_adjust.R
\name{direct_adjust}
\alias{direct_adjust}
\title{Directly age-adjusted rates}
\usage{
direct_adjust(df, agegroup, events, person_yrs, std_pop, base = 1e+05,
  level = 95)
}
\arguments{
\item{df}{A data frame.}

\item{agegroup}{Age group or other stratifying variable}

\item{events}{Number of events.}

\item{person_yrs}{Number of person-years at risk.}

\item{std_pop}{Vector of standard population. Can be totals, proportions, or percentages.}

\item{base}{Multiplier; e.g. per 100,000 population.}

\item{level}{Confidence level expressed as percentage.}
}
\value{
A data table with the following fields:
\describe{
  \item{\code{events}}{Number of events.}
  \item{\code{person_yrs}}{Total person-years at risk.}
  \item{\code{adj_rate}}{Age-adjusted rate.}
  \item{\code{adj_rate_stderr}}{Standard error of age-adjusted rate.}
  \item{\code{adj_lci}}{Lower confidence limit of age-adjusted rate per Tiwari (2006)}
  \item{\code{adj_uci}}{Upper confidence limit for age-adusted rate per Tiwari (2006)}
  \item{\code{crude_rate}}{Crude (unadjusted) rate.}
  \item{\code{crude_lci}}{Lower confidence limit for crude rate.}
  \item{\code{crude_uci}}{Upper confidence limit for crude rate.}
  }
}
\description{
Directly age-adjusted rates
}
\examples{
# US age-adjusted cancer rates by year and sex
# using standard SEER age groups 0, 1-4, 5-9, 10-14, 15-19, ..., 80-84,
library(tidyepi)
library(dplyr)
seer_pop <- c(3795, 15192, 19920, 20057, 19820, 18257,
             17722, 19511, 22180, 22479, 19806, 17224,
             13307, 10654,  9410,  8726,  7415,  4900, 4259)
cancer_by_year <- cancer \%>\%
  group_by(Year, Sex) \%>\%
  do(direct_adjust(., agegroup, n, pop, seer_pop))

}
\references{
Tiwari RC et al. (2006) Efficient interval estimation for age-
adjusted cancer rates, Statistical Methods in Medical Research 15:547-569.

Garwood F (1936) Fiducial limits for the Poisson distribution,
 Biometrika 28:437-442.
}
