% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_pop.R
\docType{data}
\name{dist_lst}
\alias{dist_lst}
\title{Standard population distributions.}
\format{A named list with 24 named vectors: master age grouping, SEER age grouping, and distributions #1 through #22 as published by NCHS}
\usage{
data(dist_lst)
}
\description{
Standard 2000 populations for direct age adjustment, from National Center for Health Statistics
}
\references{
Klein RJ, Schoenborn CA. Age adjustment using the 2000
projected U.S. population. Healthy People Statistical Notes, no. 20.
Hyattsville, Maryland: National Center for Health Statistics. January 2001.
https://www.cdc.gov/nchs/data/statnt/statnt20.pdf
}
\keyword{datasets}
