% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_pop.R
\docType{data}
\name{master_tbl}
\alias{master_tbl}
\title{STandard population master table.}
\format{A data table with 24 rows and 2 variables.
\describe{
  \item{\code{age_group}}{Age group}
  \item{\code{pop}}{Standard population, in thousands}
 }

#' @references Klein RJ, Schoenborn CA. Age adjustment using the 2000
projected U.S. population. Healthy People Statistical Notes, no. 20.
Hyattsville, Maryland: National Center for Health Statistics. January 2001.
https://www.cdc.gov/nchs/data/statnt/statnt20.pdf}
\usage{
data(master_tbl)
}
\description{
Standard 2000 master population for direct age adjustment, from National Center for Health Statistics
}
\keyword{datasets}
