% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent.R
\name{percent}
\alias{percent}
\title{Convert a number to a percentage.}
\usage{
percent(x, places = 1)
}
\arguments{
\item{x}{A number.}

\item{places}{Number of decimal places in final result (default = 1).}
}
\value{
The number expressed as a percentage.
}
\description{
Convert a number to a percentage.
}
