% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir.R
\name{reshape_for_SIR}
\alias{reshape_for_SIR}
\title{Data reshaping for indirect adjustment}
\usage{
reshape_for_SIR(df, strata, split_var, study_group, n, pop)
}
\arguments{
\item{df}{A data frame.}

\item{strata}{Variable that contains strata (such as age).}

\item{split_var}{Variable to split into study and referent groups.}

\item{study_group}{Value of split_var to use as study group.}

\item{n}{Number of events.}

\item{pop}{Number of person-years at risk.}
}
\value{
A data table with the following fields:
\describe{
  \item{\code{strata}}{Grouping variable}
  \item{\code{study_count}}{Events in study group}
  \item{\code{study_pop}}{Person-years at risk in study group}
  \item{\code{ref_count}}{Events in referent group}
  \item{\code{ref_pop}}{Person-years at risk in referent group}}
}
\description{
Reshape data to four columns for indirect adjustment.
}
