% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seer_weight.R
\docType{data}
\name{seer_weight}
\alias{seer_weight}
\title{SEER standard population weights.}
\format{A data frame with 19 rows and 3 variables:
\itemize{
  \item agegroup SEER age group 0, 1-4, 5-9, 10-14, ..., 75-84, 85+
  \item std_pop  2000 US standard population (thousands)
  \item wgt      2000 US standard population (proportions)
 }}
\usage{
data(seer_weight)
}
\description{
Dataset of standard population weights
}
\references{
Klein RJ, Schoenborn CA (2001) Age Adjustment Using
the 2000 Projected U.S. Population. Statistical Notes No. 20,
Centers for Disease Control and Prevention / National Center for Health
Statistics
\url{https://www.cdc.gov/nchs/data/statnt/statnt20.pdf}
}
\keyword{datasets}
