% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_pop.R
\docType{data}
\name{std_pop_single_age}
\alias{std_pop_single_age}
\title{Standard population distributions by single year of age.}
\format{
A data frame with two columns:
 \describe{
  \item{\code{age}}{Year of age from 0 to 100+}
  \item{\code{population}}{Standard population in thousands.}
  }
}
\source{
{US Bureau of the Census} \url{https://www.census.gov/prod/1/pop/p25-1130/p251130.pdf}
}
\usage{
data(std_pop_single_age)
}
\description{
Standard 2000 populations for direct age adjustment, from National Center for Health Statistics
}
\references{
US Bureau of the Census. Population projections of the United States by age, sex, race, and Hispanic origin: 1995 to 2050 (P25-1130). February 1996, pp. 52-53.

National Cancer Institute, Surveillance, Epidemiology, and End Results Progam. Standard populations (millions) for age-adjustment.
https://seer.cancer.gov/stdpopulations
}
\keyword{datasets}
