% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlated-ratio.R
\name{correlated_rates}
\alias{correlated_rates}
\title{Ratios of correlated age-adjusted rates.}
\usage{
correlated_rates(df, region, agegroup, events, person_yrs, std_pop,
  parent = "Parent Region", base = 1e+05, level = 95, dec_rate = 1,
  dec_ratio = 2)
}
\arguments{
\item{df}{A data frame with columns for region, age group, event counts, and person-years totals as described in the next four arguments}

\item{region}{Subregion for which to calculate adjusted rates.}

\item{agegroup}{Age group or other stratifying variable.}

\item{events}{Number of events.}

\item{person_yrs}{Number of person-years at risk.}

\item{std_pop}{Vector of standard population distribution. Can be totals, proportions, or percentages.}

\item{parent}{Name of parent region.}

\item{base}{Multiplier; e.g. per 100,000 population.}

\item{level}{Confidence level expressed as percentage.}

\item{dec_rate}{Decimal places to round adjusted rates.}

\item{dec_ratio}{Decimal places to round rate ratios.}
}
\value{
A data table with the following fields:
\describe{
  \item{\code{events}}{Number of events.}
  \item{\code{person_yrs}}{Total person-years at risk.}
  \item{\code{adj_rate}}{Age-adjusted rate.}
  \item{\code{adj_rate_stderr}}{Standard error of age-adjusted rate.}
  \item{\code{adj_lci}}{Lower confidence limit of age-adjusted rate per Tiwari (2006)}
  \item{\code{adj_uci}}{Upper confidence limit for age-adusted rate per Tiwari (2006)}
  }
}
\description{
Function to compute ratios of correlated age-adjusted rates with a confidence interval that accounts for the correlation. This is useful for comparing the rate of a subregion (e.g. a county) with its parent (e.g. its state).
}
\examples{
# Age-adjusted COPD mortality rates by state with ratio to US rate and confidence intervals, 2016
library(tidyepi)
copd_rates <- correlated_rates(copd, State, agegroup, Deaths, Population, 
              std_pop = c(23961, 18136, 12315, 4259), parent = "United States") 

}
\references{
Tiwari RC et al. (2006) Efficient interval estimation for age-adjusted 
cancer rates. Statistical Methods in Medical Research 15:547-569. 
https://www.ncbi.nlm.nih.gov/pubmed/17260923

Tiwari RC et al. (2010) Interval estimation for ratios of correlated age-adjusted rates.
Journal of Data Science 8:471-482. https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3279758/
}
