% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_adjust.R
\name{direct_adjust}
\alias{direct_adjust}
\title{Directly age-adjusted rates}
\usage{
direct_adjust(df, agegroup, events, person_yrs, std_pop, base = 1e+05,
  level = 95, places = 1)
}
\arguments{
\item{df}{A data frame.}

\item{agegroup}{Age group or other stratifying variable}

\item{events}{Number of events.}

\item{person_yrs}{Number of person-years at risk.}

\item{std_pop}{Vector of standard population. Can be totals, proportions, or percentages.}

\item{base}{Multiplier; e.g. per 100,000 population.}

\item{level}{Confidence level expressed as percentage.}

\item{places}{Decimal places to round results.}
}
\value{
A data table with the following fields:
\describe{
  \item{\code{events}}{Number of events.}
  \item{\code{person_yrs}}{Total person-years at risk.}
  \item{\code{adj_rate}}{Age-adjusted rate.}
  \item{\code{adj_rate_stderr}}{Standard error of age-adjusted rate.}
  \item{\code{adj_lci}}{Lower confidence limit of age-adjusted rate per Tiwari (2006)}
  \item{\code{adj_uci}}{Upper confidence limit for age-adusted rate per Tiwari (2006)}
  \item{\code{crude_rate}}{Crude (unadjusted) rate.}
  \item{\code{crude_lci}}{Lower confidence limit for crude rate.}
  \item{\code{crude_uci}}{Upper confidence limit for crude rate.}
  }
}
\description{
Directly age-adjusted rates
}
\note{
Confidence limits are computed using the method of Tiwari et al.
(2006). The upper limit is adjusted with a continuity correction prompted 
by the use of a continuous distribution (gamma) to approximate a discrete 
random variable (Poisson).
}
\examples{
# US age-adjusted cancer rates by year and sex
# using standard SEER age groups 0, 1-4, 5-9, 10-14, 15-19, ..., 80-84,
library(tidyepi)
library(dplyr)
cancer_by_year_sex <- cancer \%>\%
  group_by(Year, Sex) \%>\%
  do(direct_adjust(., agegroup, n, pop, std_pop_list$seer_pop))

}
\references{
Anderson RN and Rosenberg HM (1998) Age standardization of 
death rates: Implementation of the year 2000 standard. National Vital 
Statistics Reports 47(3). Hyattsville, Maryland: National Center 
for Health Statistics. https://www.cdc.gov/nchs/data/nvsr/nvsr47/nvs47_03.pdf

Tiwari RC et al. (2006) Efficient interval estimation for age-adjusted 
cancer rates. Statistical Methods in Medical Research 15:547-569. 
https://www.ncbi.nlm.nih.gov/pubmed/17260923
}
